Fix seeking dead-zone when data gets evicted.
Checks to see if a seek is within the buffered zone or not.
Clears the SourceBuffer if seeking outside the buffered zone.

This allows the efficient indexed lookup for inserted segments to still
be used while streaming, optimizes buffered seeks, keeps the playback
timeline sane in spite of the browser's eviction strategy, and obviates
the need for an earlier fix for some other MSE weirdness.

Closes #15.

Change-Id: I0226ebb679fab6739d3d0cc071e22f434983c9d2