Fix ContentDatabase test failures.
Tests were failing if an existing database had a greater version
number.  This was an artifact of an older version of ContentDatabase
which used a different version number.

Tests would also fail if one of the 'stores a stream' tests was run
several times in a row, such that the stream ID had more than one
digit.  The reference matcher now accepts multi-digit IDs.

Change-Id: I8d8949e46bd0720abd761248790b2ee5965ff4a1