Restrict test app localStorage to one key.
Because localStorage is shared between pages on the same domain,
the way the test app used an arbitrary number and set of keys
causes a conflict with the documentation, which uses localStorage
to remember preferences.  This manifests in the appearance of an
unusable MPD in the offline list called "show", since this is the
localStorage key used by the docs.

Change-Id: Ie906f9d2f3259dae33d3682ce61e942e27531d26