Fork SegmentTemplate test asset as non-live.
The test asset as it exists upstream is styled as a live stream, in
spite of the fact that it is static content.  This creates a bad
experience in the test app when this asset is selected.  Fork it and
remove the update property that makes the Player treat it as live.

Change-Id: I8bc18c68186ccf9f51e9e031753bb39ed33e065a