Change underflow system in Player.
Underflow is now possible before the playhead enters a gap in the
buffered ranges.  Some misbehaving live content freezes with as much
as 50ms of data still buffered, so going into an underflow/buffering
state with a positive amount of buffered data is desirable.

To keep this from causing trouble at EOF, there is a special check
against video.duration.

Change-Id: I5a350c543fe33910a55cd8e9ebd8fd1ab2d6408c