mpd parser: Handle minimumUpdatePeriod being set to 0.
For live streams, minimumUpdatePeriod can be set to 0. It means that
we shouldn't refresh the manifest until there's an explicit signal
from within the stream. As of now, there's no way to get that signal
from MSE. So as a workaround, if we see a minimumUpdatePeriod of 0,
we set it to 10 seconds. This is similar to how Exoplayer and Dash.js
handle this case.

[1] https://github.com/Dash-Industry-Forum/dash.js/blob/fe14ee023c0e3aa073284ce034758942e8f3c7bf/build/test/js/utils/MPDHelper.js#L12
[2] https://github.com/google/ExoPlayer/blob/953e1118a58a269ded8b0246874bd827d900fcf0/library/src/main/java/com/google/android/exoplayer/dash/DashChunkSource.java#L270