Suppress test failure in Chrome Linux
There's an intermittent test failure showing up in Chrome on Linux,
and we've traced it to a bug in Chrome.  Until the bug is fixed,
detect the ffmpeg decode failure and restart the test.  This seems
to be a reasonable workaround for now, and will be removed when the
bug in Chrome has been fixed in the channels we're using for
automated testing.

Change-Id: I1fa3fed2e83ce12f308714b44809efde5fd1e626