Fix type-inference issues with newer compilers
Newer versions of the Closure Compiler issue warnings for several
anonymous functions due to what appears to be a failure to infer
certain types.  This patch uses TypedBind to mollify the compiler.

This patch also fixes another issue raised by the newer compiler,
which is that index does not have duration and key_system properties.

b/21559591

Change-Id: Icff84ad084bfcd049f962bd4a0698b3fc78fc279