No "beep" character on linter failures
By default, the linter produces a "beep" character (ASCII 7) on
failure.  This caused an extra character to appear in the buildbot's
log on the beginning of a line that should have served as a delimiter
to make parsing easier.  The beep therefore caused a parse failure in
the buildbot when it was proofing pull-requests with linter errors.

(For an example of this failure, see PR #144.)

The solution is the linter's --nobeep flag.

Change-Id: I1a0ea7529134f9556a849fd0dd6ccb70832e6a38