Fix off-by-one error storing offline content
The final chunk was only being stored when the endTime of the final
segment happened to be null.  An index with explicit end times for
all segments, even on one stream, would trigger this bug.

For example, the end time of the final video segment may be null, but
the end time of the final audio segment may be equal to the content
duration.  The stored audio would be missing a chunk, causing the
computed stream limit to be too short on playback.

This corrects the off-by-one error in the storage code by explicitly
signalling when a segment is the last one, rather than checking for
a null endTime.

Closes #157

Change-Id: Ic99d53b9e2b071409d4d9472b8bc4ae0cf76f940