Use configure to disable cache-busting
This does not mean that Shaka v1 is suddenly cache-friendly, but this
gives users the option to weigh the problems caused by caching in
Shaka v1 against the problems caused by the cache-buster for certain
CDNs or authentication schemes.  Previously, this would have required
local changes.  Now it can be done in a standard build.

The parameter name is extra long and wordy to make sure that it cannot
be used without acknowledging that there are trade-offs involved.

Closes #235, #238, #76

Change-Id: I934a5703183d31250458b80cb38d2b5deae410df