Added configurable build framework.
Created a new build script written in python.  This can be used to
build one of several available build types as well as any custom
build.  It uses a custom file format to specify the contents of a
build.  The default is to build the complete build.

There will be a follow up change to replace the rest of the build
scripts with python.  This replaces lib.sh and build.sh, but lib.sh
still exists.

This also updates the stat script for the new output location and
now supports giving the build type name as the argument.  Also can
now be run with any working directory.

Change-Id: I1243e8ad2c7e780ef67d002d4ac8d7d5e3bd5211