Add stack-based messages to all assertions for IE
IE11 does not show stack traces on failed assertions in the console.
So if there is no message associated with a failed assertion, create
one based on the stack trace.  This should help debug errors on IE.

Note that in Shaka 2, we already require messages for all assertions,
so this is just a stop-gap measure to fix lingering bugs in Shaka 1.

Issue #251

Change-Id: Ib42c693278485fa2d3bad7e34f3c73d15d23be27