Fix (some) test failures on IE11
This fixes some of the (numerous) test failures on IE11:

- Move SBM assertion after spurious event check.
- Install polyfills before running all tests.
- Fix missing closing tag (caused SyntaxError in IE11's XML parser).
- Avoid Uint8Array.slice, not present on IE11.
- Remove 'done' argument from describe().
- Make interpretContentProtection use childNodes instead of children.
- Errors only show stack trace if thrown.
- Extend some timeouts to reduce flake.
- Match different error type for DataView errors.

Tests are still not clean on IE, and many make Chrome-specific
assumptions, but this is at least an improvement.  Shaka 2's tests are
clean on IE, so it may not be worth the effort to fix all tests on IE
in Shaka 1.

Working toward a fix for #251

Change-Id: I156024f9879fa35893886b3181a526483f0dd947