Improve video unload mechanism
It was pointed out in #80 that setting video.src='', although common,
is not best practice.  I finally received some advice from
w3c/media-source#53, specifically a pointer to the section "Best
practices for authors using media elements" in the HTML5 spec.
The solution is to use both removeAttribute('src') and load().
As determined during the investigation of #80, removeAttribute('src')
alone is not effective.

Change-Id: If0eb0dd2dc60dee0673dc92dabe9dc3e913457db