Fix timeline duration for multi-period live streams
Multi-period live streams have no presentation duration and have no duration
on the final period.

In this case, we were setting the timeline duration to the start time of the
final period, based on end = start + duration with duration == null.

This fixes the situation so that timeline duration is infinite.

Relates to #310

Change-Id: I8592318e3d4c5643fa1fcc03b34624957a2e974c