Fix stream selection for multicodec assets
When playing content with both WebM and MP4, we can choose whichever
the browser supports.  If the browser supports both, the choice is
arbitrary.

If the browser supports both, however, this does not mean we can
switch between them.  So once a choice has been made, filter out
streams which are incompatible.  If we're playing WebM, stick to it.

Change-Id: I03d83627cb9ce14e8018d540d2ecdb30eba09600