Remove invalid assertion.
If a user seeks while an update is in progress,
waitingToClearBuffer can be set to true. When that update
finishes, performingUpdate will be set to false. So, if the user
seeks again before onUpdate_ is called, waitingToClearBuffer
can be true and performingUpdate will be false, thus, the
assertion in seeked() is not valid.

Closes #334

Change-Id: Id6517078bf0e9037ead2d357c4db241f381a383c