Fix Player deferred switches WRT clear buffer.
Now when using deferred switches in Player, it will track whether
to clear the buffer.  This is needed because during startup,
StreamingEngine will append the first media segment; the old code
assumes that the buffer is empty.  This is especially important for
text tracks which may only be one segment.

Change-Id: I03d97e2a674c8e713145dbc619b15d7c9402fb7f