Revert "Update clock sync on manifest update"
Hitting UTCTiming sources over and over could be considered abusive
behavior at scale, so we will only sync the clock once on startup.

The change to update clock sync on manifest update was added for #345,
but it seems that this was not part of the fix, since the content in
that issue had no UTCTiming element at all.

Change-Id: I32f4fc0a40fc92feeb54191e5fe869e0aaa9b54c