Fix crash in TextEngine when text is not visible
Setting TextTrack mode to 'disabled' causes cues to go null.  It
should not be used on an active track.  Instead, we should use
'hidden' for an active track that isn't being displayed.  We will
still use 'disabled' for pre-existing tracks that are not being used
by Shaka.

Also corrects detection of existing Shaka-created TextTracks.
(Our ID was actually a label.)

https://www.w3.org/TR/html5/embedded-content-0.html#text-track-mode

Reported in PR #354

Change-Id: I3b8d06da7374e4a0c33fcec11bbee5c03b21941c