Change the offline UI for demo app.
Now the demo assets can only be stored once, if an asset is stored
the store button will disable.  Combined the store and delete buttons
into one.  Added a text box to set the name of the asset when
storing custom assets.  Also add export to Storage.destroy so it
works in the compiled version.

b/29777213

Change-Id: I7972a7931166c67efcffe70f5017a920cca617ef