Attempt to fetch one chunk prior on seeking
As per issue #330, manifests are only required to be approx.
correct with respect to their description of the media segments.
In order to avoid stalling when seeking through content like this,
fetch the chunk prior to the one that supposedly contains the
current time.

This is an attempt to apply a similar fix to Ia7bb74d0120.