Fix play controls state on Android
Since Android has disabled autoplay, we need to make it clear that the
user must click the play button.  To do that, the controls need to
know when to recalculate play state once the video has loaded.  The
events we were using for this seem not to be reliable.  Now the
application explicitly signals the controls for this.

Closes #432

Change-Id: Ic6bfda9c32ae7326b2e632d3cc3015a921dd9059