Correct usage of jasmine's toBeCloseTo matcher
The second argument is meant to be the number of decimal places to
which the value must match, not the maximum difference.  So using
"0.1 decimal places" meant the maximum difference was actually 1.25.
You must pass "1" to get a maximum difference of "0.1".

This also adjusts some of the expectations to fix affected tests.

Change-Id: Ibc1223297011063e3e4f3fff0ba64fde06292128