Fix bug with text engine buffered ranges.
Based on pull request by @baconz.  Before, when appending segments,
TextEngine was given the segment times and then added the timestamp
offset to get the real time.  However, this is incorrect as we want
the time relative to the presentation timeline.  So we should only be
adding the period start time (the timestamp offset also includes PTO).

Closes #411

Change-Id: I92582d1cda034daef5e8ce2d7b79deafa0ddf828