Use buffer end to determine which segments to download.
Before, when we adapt we tell StreamingEngine we are in an unbuffered
state.  This causes it to download segments at the playhead, even
though we have some buffered.  Normally this does not cause problems,
but on Edge, it can cause skipping.  So now, StreamingEngine will
use buffer end to determine which segments to download.

Issue #446

Change-Id: Ib1f4cd3a1627efa957825eb0445c99f753f31e7b