Use --dependency_mode=LOOSE for library compiles
This change does not order input files for the --checks_only
pass used to type-check the unit tests. That is,
--dependency_mode=NONE (the default) is specified for testing
code. This works as the alphabetical order of source files from
build/all.py happens to not result in any errors. However,
as files are renamed in the future this could change.