Update YouTube assets and callbacks
We were failing to make license requests to the YouTube license server
for PlayReady.  This updates the YouTube callbacks to correct this by
dropping PlayReady headers from the request.

Also fixes an IE 11 compatibility issue that was covered up by the
failing license request: startsWith() is not available on IE 11.

Note that the affected YouTube assets still fail to play on IE and
Edge due to another bug we have yet to identify.

Change-Id: Ibc1c9d685adba0caa163902bb3de5741c67efa83