Player.configure takes plain Object
The reality is that it has always taken a plain Object, but we were
calling it shakaExtern.PlayerConfiguration.  While it is true that
the argument follows that layout, only the fields to be changed are
needed.  This becomes an important distinction for other Closure-based
projects using our compiled binary.

As we move toward generating externs for the library at compile-time,
we should be careful about the required input type so our generated
externs can be used directly by the app.

As a side-benefit, this also removes some casts from our tests.

Change-Id: Iadb6a071d502e9ac6955af4d85af918eedd23571