Reduce the default buffering goal
Since we no longer clear ahead of the playhead, the only way to assure
responsive visual upgrade is not to buffer too far ahead.  Therefore
our new default buffering goal is 10 seconds instead of 30 seconds.

Manual testing using Chrome's network throttling feature has shown
no increase in buffering rates, so 10 seconds seems to be enough for
stable connections.

For situations in which 10 seconds is not enough, the bufferingGoal
setting can still be customized by the application.

Closes #520

Change-Id: I248c1c93dd4cdd037e6e3ca12bb207ba54845836