Fix buffering watchdog to match StreamingEngine
StreamingEngine continues to buffer until the amount buffered is
greater than or equal to the goal.  Playhead's watchdog checked that
the amount buffered was strictly greater than the goal, leading to a
race in which we could get stuck buffering forever.

The race was between buffering and the first watchdog timer callback.

If data gets buffered before the first callback fires, the watchdog
will see that bufferedAhead is greater than threshold, and it will
not enter a buffering state.

If the callback fires before anything is buffered, it sees
bufferedAhead == 0 and enters a buffering state.  If the config is
such that bufferingGoal and rebufferingGoal are the same, then
Playhead is stuck in buffering.  StreamingEngine won't buffer more
until something plays, and Playhead won't leave the buffering state
until something more is buffered.

This turned out to be the root cause of failures in one of our
quarantined tests.  While debugging this test, I noticed that
video.playbackRate was 0, which we set while buffering.  This led me
to investigate buffering logic in Playhead, where I noticed that
bufferedAhead was exactly rebufferingGoal in the instances where the
test failed.

In 100 test runs, the test failed 9 times.  No other tests had to be
run to trigger the failure, so this one test was run in a for loop.

Once I understood the cause of the race, I found that I could trigger
the failure in 100 out of 100 tests by changing startWatchdogTimer_()
to onWatchdogTimer_() in the Playhead constructor.  This ensures that
the timer callback fires synchronously, before anything can be
buffered.

With the underlying bug fixed, the test has now been removed from
quarantine.

Change-Id: Ief5c11b2211045a32dc0a9f711b17a4ec1bf55af