Fix unload/load bug in Player.
Based on PR #613

There was a bug where calling unload() right before calling load()
would cause a race that would sometimes cause a failure.  This is
because the fields are reset before the unload is complete so the
second call to unload() (from inside load) will complete immediately.

So now we store the unload Promise chain while it is in progress so
we can wait on it from load().

Closes #612

Change-Id: I6c0cdd931827d709fc41322edd51fe10e4aa87ae