Update Edge workaround in DrmEngine.
Edge recently pushed an update which fixed a bug that we were working
around.  This was that it didn't support multiple MIME types when
querying EME.  Now it will allow multiples to be passed in.

However, Edge's update also introduced a new bug that when getting the
configuration from the access object, it doesn't report the correct
MIME types.  It should include every MIME type that is supported,
but it only includes the first one.

Closes #634

Change-Id: I7156887a112dc844b3cbc9bb2dd9676afee6cf56