Use the release compiled version in the demo.
There are two versions of the compiled library: a debug version and a
release version.  A recent change altered the differences. Before, the
only difference was that the debug had the source mapping. Now the
debug version doesn't rename anything and exposes all internal members.

To get the source mapping, the demo used the debug version.  Now both
versions have the source map, so we should be using the release
version of the compiled library in the demo.

Change-Id: I33b4af6532e0eaf0749ffd8341d4f9d72a1ff0fa