Increase Playhead fudge factor for IE/Edge.
IE/Edge have problems with reporting the buffered ranges when we get
to the end of the media.  We use a fudge factor to determine if we have
buffered all the content.  It was too small which meant that we would
not detect end of stream correctly.  This increases the fudge factor
so we detect it correctly.

Closes #658

Change-Id: Ie9ecf8fd49eb716ffd8ccd9bcdff53b88356bbfe