Whitelist svg camelcase attributes in htmlhint
This is better than ignoring all camelcase.

Introduced by yaniswang/HTMLHint#121 and released in 0.9.13.

However, 0.9.13 is incompatible with node v0.10.  This was fixed
in yaniswang/HTMLHint#132, but there has not been a release since
(in 8 months).

Therefore we are pinning to today's git version instead of a release.
This gets us the camelcase whitelisting feature and doesn't break with
node v0.10.

Change-Id: I1e3b6c0c96ef24662b46497b1a601bde0dae5704