Don't use |session.closed| in DrmEngine.destroy.
If a session has been created but the license request hasn't been sent,
the session is not "callable".  This will cause the |session.close()|
to be rejected and the |session.closed| Promise to never resolve.  So
in DrmEngine.destroy, we should be using the Promise returned from
|session.close()| instead of |session.closed|.  The |session.close()|
Promise will only resolve once the |session.closed| Promise is resolved
so they are effectively the same.

Closes #664

Change-Id: Ia619740d1eb3fa8ca1bdf121574d1efb1ddbea26