Use Playhead to determine when changing Periods.
Before, we would fire a 'trackschanged' event when StreamingEngine
asked us to select streams for a new Period.  However, this is not
when the tracks change, they change when the playhead moves into a
different Period.

This adds code to Playhead to track where the current Period is and
raise an event when it changes.  This also reduces any extra events so
a 'trackschanged' event is only fired when the tracks actually change.

Issue #680

Change-Id: I85aedaed737f9324d5871201d952066886ea290c