Track active tracks in Player.
Before, the Player would query StreamingEngine for the "active" streams.
The Player would use this to set the |track.active| field.  However,
this causes problems with multi-Period content.  Once StreamingEngine
has buffered through a Period transition, it will report the active
streams from the new Period.  This will mean that in
|player.getVariantTracks|, none of the tracks will be active.

Now, Player will track the IDs of the active tracks for every Period
so it can correctly report which tracks are active.

Closes #680

Change-Id: Ib5ffc88b32e165f088e6472930eeb1b29b5adfe9